refactor column mapper
